refactor: CreditCardRegisterDone 구조 개선
